Organize subscription files and move global style
