+++ web/components/admin-main/admin.html
@@ -6,7 +6,7 @@
   <title>Amul Watchdog Control</title>
   <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap" rel="stylesheet">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="../style/style.css">
+  <link rel="stylesheet" href="../../style.css">
 </head>
 <body>
   <script>
