+++ web/components/subscription/subscriptions-ui.js
@@ -3,7 +3,7 @@
 // subscription-modal.js and subscription-helpers.js respectively.
 // The fetchAPI utility has been moved to utils.js.
 
-import { initSubscriptionsUI } from '../subscription-modal/subscription-modal.js';
+import { initSubscriptionsUI } from './subscription-modal.js';
 
 // Initialize the subscriptions UI (which now primarily means the modal)
 initSubscriptionsUI();
