+++ web/components/user-main/user.html
@@ -6,7 +6,7 @@
   <title>User Dashboard</title>
   <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap" rel="stylesheet">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="../style/style.css">
+  <link rel="stylesheet" href="../../style.css">
 </head>
 <body>
   <div id="particles-js" style="position: fixed; width: 100%; height: 100%; z-index: -1; top: 0; left: 0;"></div>
